pyproject.toml
[project]
name = "ai-tutor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.26.0",
    "google-genai",
    "google-cloud-aiplatform",
    "pillow",
    "python-dotenv",
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "requests",
    "PyJWT",
    "gradio",
    "litellm",
    "json5",
    "playwright",
    "volcengine",
    "volcengine-python-sdk[ark]",
    "pydantic[email]",
    "bcrypt",
    "trafilatura",
    "tavily",
    "stripe",
    "flask",
    "aiosqlite",
    "alibabacloud_imageseg20191230",
    "pdfdeal",
    "opencv-python",
    "markitdown[all]",
    "defusedxml",
    "pypandoc",
    "python-pptx",
    "manim",
    "matplotlib",
    "imageio",
    "imageio-ffmpeg",
    "numpy",
    "manim-voiceover",
    "manim-voiceover-qwen3-tts",
    "kokoro-manim-voiceover",
    "dashscope",
    "httpx[socks]",
    "requests-toolbelt",
    "openai",
    "asyncddgs",
    "tavily-python",
    "sqlalchemy",
    "pytest",
    "pytest-mock",
    "loguru",
]
